enforce implementation of repeat_pixel_to_buffer in Model trait
