+++ mipidsi/src/models/ili9342c.rs
@@ -47,6 +47,10 @@ impl Model for ILI9342CRgb565 {
     {
         ili934x::write_pixels_rgb565(dcs, colors)
     }
+
+    fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
+        crate::graphics::repeat_pixel_to_buffer_rgb565(color, buf)
+    }
 }
 
 impl Model for ILI9342CRgb666 {
@@ -83,6 +87,6 @@ impl Model for ILI9342CRgb666 {
     }
 
     fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        super::repeat_pixel_to_buffer_rgb666(color, buf)
+        crate::graphics::repeat_pixel_to_buffer_rgb666(color, buf)
     }
 }
