+++ mipidsi/src/models/ili9486.rs
@@ -61,7 +61,7 @@ impl Model for ILI9486Rgb565 {
     }
 
     fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        super::repeat_pixel_to_buffer_rgb565(color, buf)
+        crate::graphics::repeat_pixel_to_buffer_rgb565(color, buf)
     }
 }
 
@@ -110,7 +110,7 @@ impl Model for ILI9486Rgb666 {
     }
 
     fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        super::repeat_pixel_to_buffer_rgb666(color, buf)
+        crate::graphics::repeat_pixel_to_buffer_rgb666(color, buf)
     }
 }
