+++ mipidsi/src/models/st7735s.rs
@@ -91,6 +91,6 @@ impl Model for ST7735s {
     }
 
     fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        super::repeat_pixel_to_buffer_rgb565(color, buf)
+        crate::graphics::repeat_pixel_to_buffer_rgb565(color, buf)
     }
 }
